% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_regression.R
\name{lm_regression}
\alias{lm_regression}
\title{A function to generate the prediction values for random forest regression.}
\usage{
lm_regression(y, X_train, X_test, ...)
}
\arguments{
\item{y}{a \code{numeric} that contains the real valued response variable.}

\item{X_train}{a \code{matrix} that contains the features for tarining set, nrow(X) = length(y).}

\item{X_test}{a \code{matrix} that contains the features for testing set.}

\item{...}{additional arguments passed to \code{randomForest} defined in the package \code{randomForest}.}
}
\value{
a \code{vector} that containing the predictions by linear regression.
}
\description{
A function to generate the prediction values for random forest regression.
}
\details{
\code{lm_regression} conduct linear regression.
}
