% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomForest_f.R
\name{randomForest_f}
\alias{randomForest_f}
\title{A function to generate the svm decision values.}
\usage{
randomForest_f(y, X_train, X_test, ...)
}
\arguments{
\item{y}{a \code{factor} that contains the binary response variable with levels 1,0 for prediction.}

\item{X_train}{a \code{matrix} that contains the features for tarining set, nrow(X) = length(y).}

\item{X_test}{a \code{matrix} that contains the features for testing set.}

\item{...}{additional arguments passed to \code{randomForest} defined in the package \code{randomForest}.}
}
\value{
a \code{vector} that containing the positive voting proportions of the random forest model for testing set.
}
\details{
\code{svm_f} conduct binary classification using randomForest.
}
