% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_f.R
\name{svm_f}
\alias{svm_f}
\title{A function to generate the svm decision values.}
\usage{
svm_f(y, X_train, X_test, ...)
}
\arguments{
\item{y}{a \code{factor} that contains the binary response variable with levels 1,0 for prediction.}

\item{X_train}{a \code{matrix} that contains the features for tarining set, nrow(X) = length(y).}

\item{X_test}{a \code{matrix} that contains the features for testing set.}

\item{...}{additional arguments passed to \code{svm} defined in the package \code{e1071}.}
}
\value{
a \code{vector} that is the same length of y containing the decision values on testing set.
}
\details{
\code{svm_f} conduct binary classification using SVM.
}
