% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_regression.R
\name{svm_regression}
\alias{svm_regression}
\title{A function to generate the prediction values for svm regression.}
\usage{
svm_regression(y, X_train, X_test, ...)
}
\arguments{
\item{y}{a \code{numeric} that contains the real valued response variable.}

\item{X_train}{a \code{matrix} that contains the features for tarining set, nrow(X) = length(y).}

\item{X_test}{a \code{matrix} that contains the features for testing set.}

\item{...}{additional arguments passed to \code{svm} defined in the package \code{e1071}.}
}
\value{
a \code{vector} that is the same length of y containing the prediction values on testing set.
}
\description{
A function to generate the prediction values for svm regression.
}
\details{
\code{svm_regression} conduct support vector regression.
}
