% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_cv.R
\name{predict_cv}
\alias{predict_cv}
\title{A function to generate the cross validated prediction/decision values.}
\usage{
predict_cv(y, X, pred_f = svm_class, k = 10, p = 1, ...)
}
\arguments{
\item{y}{an object that contains the binary response factor with levels 1,0 for binary classification, or a numeric vector for the prediction problem.}

\item{X}{a \code{matrix} that contains the features, nrow(X) = length(y).}

\item{pred_f}{a function that gets the input argument named y , X_train, and X_test, and output is the real numbered decision values for X_test.

For example, the returned values can be the votes proportions for random forest, and it can also be the decision values for SVM.

This package contains buildin predictor functions for you to use, such as: \code{\link{svm_class}} and \code{\link{randomForest_class}}
However, it is still recommended to create your own function with tunned parameters.}

\item{k}{the number of folds used in cross validation, default 10.}

\item{p}{the number of parallels used, default equals 1, recommend to use p = k on server.}

\item{...}{additional arguments passed to \code{pred_f}}
}
\value{
a \code{vector} that is the same length of y containing the real value decision values after cross validation.
}
\description{
A function to generate the cross validated prediction/decision values.
}
\details{
\code{predict_cv} conduct predictions of machine learning binary classifier with cross validation.
This is a convenient function to get a neat and efficient result of prediction after cross validation. The output can be piped directly into the downstream performance evaluation.

The missing values will not be imputed by default, and the dimensions containing missing values will be dropped.
}
